removed unnecessary methods. Fixes issue-1
