--- src/main/java/com/dnastack/beacon/LiftOvers/LiftOverException.java
@@ -1,12 +1,0 @@
-package com.dnastack.beacon.LiftOvers;
-
-/**
- * Created by patrickmagee on 2016-05-26.
- */
-@SuppressWarnings("WeakerAccess")
-public class LiftOverException extends Exception {
-
-    public LiftOverException(String msg){
-        super(msg);
-    }
-}
