+++ src/main/java/com/dnastack/beacon/liftover/ucsc/UCSChainFile.java
@@ -1,5 +1,31 @@
-package com.dnastack.beacon.LiftOvers;
+/*
+ * The MIT License
+ *
+ * Copyright 2014 DNAstack.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.dnastack.beacon.liftover.ucsc;
 
+
+import com.dnastack.beacon.liftover.util.GenomeBuild;
+import com.dnastack.beacon.liftover.api.ChainFile;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -11,17 +37,13 @@ import java.util.zip.GZIPInputStream;
 /**
  * Class for wrapping the chain files used in the liftOver process. Extends java.io.file
  * to easily retrieve chainFiles with minimal user input
- * <p>
- * Created by patrickmagee on 2016-05-26.
+ *
+ * @author patrickmagee
  */
-public class ChainFile extends File {
+public class UCSChainFile extends ChainFile {
 
     private static final String FILE_PATH_TEMPLATE = "chains/%s/hg%dToHg%d.over.chain";
     private static final String UCSC_REMOTE_TEMPLATE = "http://hgdownload.cse.ucsc.edu/goldenPath/%s/liftOver/%s";
-
-    private final String buildFrom;
-
-    private final String buildTo;
 
     /**
      * Constructor for The Chainfile.
@@ -30,11 +52,8 @@ public class ChainFile extends File {
      * @param to   GenomeBuild that is the target
      * @throws IOException
      */
-    public ChainFile(GenomeBuild from, GenomeBuild to) throws IOException {
-        super(getChainFilePath(from, to));
-
-        buildFrom = from.getBuildName();
-        buildTo = to.getBuildName();
+    public UCSChainFile(GenomeBuild from, GenomeBuild to) throws IOException {
+        super(getChainFilePath(from, to), from, to);
 
         if (!this.exists()) {
             throw new IOException("File does not exist");
@@ -50,16 +69,11 @@ public class ChainFile extends File {
      * @param to       Target GenomeBUild
      * @throws IOException
      */
-    public ChainFile(String pathname, GenomeBuild from, GenomeBuild to) throws IOException {
-        super(pathname);
-
+    public UCSChainFile(String pathname, GenomeBuild from, GenomeBuild to) throws IOException {
+        super(pathname, from, to);
         if (from == null || to == null) {
             throw new IllegalArgumentException("GenomeBuilds cannot be null");
         }
-
-        buildFrom = from.getBuildName();
-        buildTo = to.getBuildName();
-
         if (!this.exists()) {
             throw new IOException("File Does not exist");
         }
@@ -74,15 +88,11 @@ public class ChainFile extends File {
      * @param buildTo   String of target genome build
      * @throws IOException
      */
-    public ChainFile(String pathname, String buildFrom, String buildTo) throws IOException {
-        super(pathname);
+    public UCSChainFile(String pathname, String buildFrom, String buildTo) throws IOException {
+        super(pathname,buildFrom,buildTo);
         if (buildFrom == null || buildTo == null) {
             throw new IllegalArgumentException("Build versions cannot null");
         }
-
-        this.buildFrom = buildFrom;
-        this.buildTo = buildTo;
-
         if (!this.exists()) {
             throw new IOException("File Does not exist");
         }
@@ -128,7 +138,7 @@ public class ChainFile extends File {
      */
     public static ChainFile fromFile(File file, String buildFrom, String buildTo) throws IOException {
         String pathname = file.getAbsolutePath();
-        return new ChainFile(pathname, buildFrom, buildTo);
+        return new UCSChainFile(pathname, buildFrom, buildTo);
     }
 
     /**
@@ -172,7 +182,7 @@ public class ChainFile extends File {
      * @return ChainFile
      * @throws IOException
      */
-    public static ChainFile fromUrl(URL url, String buildFrom, String buildTo) throws IOException {
+    private static ChainFile fromUrl(URL url, String buildFrom, String buildTo) throws IOException {
         if (url == null) {
             throw new IllegalArgumentException("url cannot be null");
         }
@@ -196,116 +206,6 @@ public class ChainFile extends File {
         }
         fileOutputStream.close();
         return fromFile(temp, buildFrom, buildTo);
-    }
-
-    /**
-     * Get the starting genomebuild version, if the build is supported
-     *
-     * @return GenomeBuild Version
-     */
-    public String getBuildFrom() {
-        return buildFrom;
-    }
-
-
-    /**
-     * Get the target genomebuild version, if the build is supported
-     *
-     * @return GenomeBuild Version
-     */
-    public String getBuildTo() {
-        return buildTo;
-    }
-
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg17ToHg18() throws IOException {
-        return new ChainFile(GenomeBuild.HG17, GenomeBuild.HG18);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg17ToHg19() throws IOException {
-        return new ChainFile(GenomeBuild.HG17, GenomeBuild.HG19);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg18ToHg17() throws IOException {
-        return new ChainFile(GenomeBuild.HG18, GenomeBuild.HG17);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg18ToHg19() throws IOException {
-        return new ChainFile(GenomeBuild.HG18, GenomeBuild.HG19);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg18ToHg38() throws IOException {
-        return new ChainFile(GenomeBuild.HG18, GenomeBuild.HG38);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg19ToHg18() throws IOException {
-        return new ChainFile(GenomeBuild.HG19, GenomeBuild.HG18);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg19ToHg38() throws IOException {
-        return new ChainFile(GenomeBuild.HG19, GenomeBuild.HG38);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg19ToHg17() throws IOException {
-        return new ChainFile(GenomeBuild.HG19, GenomeBuild.HG17);
-    }
-
-    /**
-     * Convienence method for retrieving a resource chain file
-     *
-     * @return Target ChainFile from resource bundle
-     * @throws IOException
-     */
-    public static ChainFile hg38ToHg19() throws IOException {
-        return new ChainFile(GenomeBuild.HG38, GenomeBuild.HG19);
     }
 
 }
