+++ src/main/java/com/dnastack/beacon/liftover/util/GenomeBuild.java
@@ -1,0 +1,49 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2014 DNAstack.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package com.dnastack.beacon.liftover.util;
+
+/**
+ * Genome Builds which have supporting data in the resource bundles
+ *
+ * @author patrickmagee
+ */
+public enum GenomeBuild {
+    HG17("hg17", 17), HG18("hg18", 18), HG19("hg19", 19), HG38("hg38", 38);
+
+    private final int build;
+    private final String buildName;
+
+    GenomeBuild(String buildName, int build) {
+        this.buildName = buildName;
+        this.build = build;
+    }
+
+    public int getBuild() {
+        return build;
+    }
+
+    public String getBuildName() {
+        return buildName;
+    }
+}
